Extract common triple-based analyzer code into a superclass
See https://gitlab.com/oersi/oersi-etl/-/issues/238
